feat:voice emotion added

+++ app.py
@@ -7,6 +7,12 @@ import tensorflow
 from flask_cors import CORS
 
 import time
+import subprocess
+
+from keras.models import model_from_json
+import librosa
+import numpy as np
+import pandas as pd
 
 
 app = Flask(__name__)
@@ -73,16 +79,70 @@ def uploader_voice_file():
         path = os.path.realpath(f.filename)
 
         mysp=__import__("my-voice-analysis")
-        result = mysp.myspsyl(f.filename, path)
-        os.remove(path)
+        result = mysp.mysptotal(f.filename, path)
+        print(result)
+        # command = "ffmpeg -i {} -ab 160k -ac 2 -ar 44100 -vn {}".format(f.filename, os.path.join('/root/server/', 'changed.wav'))
+        # subprocess.call(command, shell=True)
+        # path = os.path.realpath("changed.wav")
+
+        # result = mysp.myspsr("changed.wav", path)
+
+        # os.remove(path)
 
     return str(result)
 
 
-@app.route('/userLogin', methods = ['POST'])
-def userLogin():
-    user = request.get_json()#json 데이터를 받아옴
-    return jsonify(user)# 받아온 데이터를 다시 전송
+@app.route('/emotion_analyzer', methods = ['POST'])
+def voice():
+    if request.method == 'POST':
+        try:
+            f = request.files['voice_file']
+        except:
+            return 'File is missing', 404
+
+        f.save(secure_filename(f.filename))
+        path = os.path.realpath(f.filename)
+
+        json_file = open('voice_model.json', 'r')
+        loaded_model_json = json_file.read()
+        json_file.close()
+        loaded_model = model_from_json(loaded_model_json)
+        # load weights into new model
+        loaded_model.load_weights("Emotion_Voice_Detection_Model.h5")
+        print("Loaded model from disk")
+
+        data, sampling_rate = librosa.load(f.filename)
+        X, sample_rate = librosa.load(f.filename, res_type='kaiser_fast',duration=2.5,sr=22050*2,offset=0.5)
+        sample_rate = np.array(sample_rate)
+        mfccs = np.mean(librosa.feature.mfcc(y=X, sr=sample_rate, n_mfcc=13),axis=0)
+        featurelive = mfccs
+        livedf2 = featurelive
+
+        livedf2= pd.DataFrame(data=livedf2)
+        livedf2 = livedf2.stack().to_frame().T
+        twodim= np.expand_dims(livedf2, axis=2)
+        livepreds = loaded_model.predict(twodim, 
+                                batch_size=32, 
+                                verbose=1)
+        prediction=livepreds.argmax(axis=1)
+        print(livepreds)
+
+        #label-list
+        # 0 - female_angry
+        # 1 - female_calm
+        # 2 - female_fearful
+        # 3 - female_happy
+        # 4 - female_sad
+        # 5 - male_angry
+        # 6 - male_calm
+        # 7 - male_fearful
+        # 8 - male_happy
+        # 9 - male_sad
+
+        os.remove(path)
+
+        return str(prediction)
+
 
 if __name__ == '__main__':
     app.run('0.0.0.0', port = 80, debug=True)
